// wwwsrc/app/hema/api/models/fft/jobs/history/ISafety.ts
export interface ISafety {
    applianceSafe: string;
    flueSafe: string;
    ventilationSafe: string;
    installationSafe: string ;
    installationTightnessTestSafe: string;
    actionCode: string;
    date: string;
    noticeStatus: string;
    noticeType: string;
    progress: string;
    report: string;
}